Support local web server
The hybrid app framework cordova uses a local web server to serve the app content to the user.

This fact requires fetching resources from a local web server. Since this is different from a usual web server request a new js module was created.

While I was at it the dependency to axios was removed as well, because it was too much for a simple web request.